Make header and footer no longer sticky
By resizing the browser viewpoint and reducing the height under a certain
threshold, the header and the footer start to overlap with the content.

Instead of using sticky headers and footers we'll extend the content container
to be at least as high as the viewport for wide screens and place the header
and footer absolute in there.